pyproject.toml
[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"

[project]
name = "kospel-cmi-lib"
version = "0.1.0a2"
description = "Python client for Kospel C.MI electric heater HTTP API"
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE*"]
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "pyyaml>=6.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
    "Typing :: Typed",
]
keywords = ["kospel", "heater", "cmi", "home-assistant", "iot", "async"]
urls = { Repository = "https://github.com/JanKrl/kospel-cmi-lib" }


[tool.uv.build-backend]
module-name = "kospel_cmi"
module-root = "src"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
# Add src to path so kospel_cmi is importable (tests mirror user imports)
pythonpath = ["src"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover"]

[dependency-groups]
dev = [
    "aioresponses>=0.7.8",
    "coverage[toml]>=7.13.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pyyaml>=6.0.3",
]
